backend/requirements.txt
fastapi
uvicorn
chromadb
langchain
langchain-community
langchain-google-genai
langchain-ollama
watchdog
python-multipart
websockets
pydantic
python-dotenv
psutil
slowapi
cryptography
pydantic-settings
google-genai>=1.0.0
google-cloud-aiplatform>=1.38.0
